requirements.txt
faster-whisper==1.0.3
requests
PyYAML
mutagen
faster-whisper
ctranslate2
transformers